Refactor component and state handling
Better separation of concerns. It’s now easier to think about the
component as having both state and a view that represents the current
state. The component’s state can be manipulated, which in turn
transforms the view to match the new state. All state changes are
captured as snapshots and thus can be reverted at will.

The example has been updated to reflect all of this.