Fix an issue setting values for attrs like `class`
The value of such attributes will be an array. This was causing an issue
in some code that was checking for nested instruction sets. To fix, we
not only check that the value is an array, but also check for a nested
array (which implies that it is, in fact, an instruction set).